+++ assemblers/http_matcher.go
@@ -17,8 +17,8 @@ type entry struct {
 	responseTimestamp time.Time
 }
 
-func newRequestResponseMatcher() httpMatcher {
-	return httpMatcher{
+func newRequestResponseMatcher() *httpMatcher {
+	return &httpMatcher{
 		messages: &sync.Map{},
 	}
 }