+++ config/config.go
@@ -9,7 +9,8 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-const timeout time.Duration = time.Second * 30
+// TODO hard-coded for now, make configurable
+const DebugAddr = "0.0.0.0:6060"
 
 var maxcount = flag.Int("c", -1, "Only grab this many packets, then exit")
 var statsevery = flag.Int("stats", 1000, "Output statistics every N packets")
@@ -26,59 +27,70 @@ var quiet = flag.Bool("quiet", false, "Be quiet regarding errors")
 // capture
 var iface = flag.String("i", "any", "Interface to read packets from")
 var fname = flag.String("r", "", "Filename to read from, overrides -i")
-var snaplen = flag.Int("s", 65536, "Snap length (number of bytes max to read per packet")
+var snaplen = flag.Int("s", 262144, "Snap length (number of bytes max to read per packet") // 262144 is the default snaplen for tcpdump
 var tstype = flag.String("timestamp_type", "", "Type of timestamps to use")
 var promisc = flag.Bool("promisc", true, "Set promiscuous mode")
 var packetSource = flag.String("source", "pcap", "Packet source (defaults to pcap)")
 var bpfFilter = flag.String("filter", "tcp", "BPF filter")
 var channelBufferSize = flag.Int("channel_buffer_size", 1000, "Channel buffer size (defaults to 1000)")
+var streamFlushTimeout = flag.Int("stream_flush_timeout", 10, "Stream flush timeout in seconds (defaults to 10)")
+var streamCloseTimeout = flag.Int("stream_close_timeout", 90, "Stream close timeout in seconds (defaults to 90)")
+var maxBufferedPagesTotal = flag.Int("gopacket_pages", 150_000, "Maximum number of TCP reassembly pages to allocate per interface")
+var maxBufferedPagesPerConnection = flag.Int("gopacket_per_conn", 4000, "Maximum number of TCP reassembly pages per connection")
 
 type Config struct {
-	Maxcount          int
-	Statsevery        int
-	Lazy              bool
-	Nodefrag          bool
-	Checksum          bool
-	Nooptcheck        bool
-	Ignorefsmerr      bool
-	Allowmissinginit  bool
-	Verbose           bool
-	Debug             bool
-	Quiet             bool
-	Interface         string
-	FileName          string
-	Snaplen           int
-	TsType            string
-	Promiscuous       bool
-	CloseTimeout      time.Duration
-	Timeout           time.Duration
-	PacketSource      string
-	BpfFilter         string
-	ChannelBufferSize int
+	Maxcount                      int
+	Statsevery                    int
+	Lazy                          bool
+	Nodefrag                      bool
+	Checksum                      bool
+	Nooptcheck                    bool
+	Ignorefsmerr                  bool
+	Allowmissinginit              bool
+	Verbose                       bool
+	Debug                         bool
+	Quiet                         bool
+	Interface                     string
+	FileName                      string
+	Snaplen                       int
+	TsType                        string
+	Promiscuous                   bool
+	StreamFlushTimeout            time.Duration
+	StreamCloseTimeout            time.Duration
+	PacketSource                  string
+	BpfFilter                     string
+	ChannelBufferSize             int
+	MaxBufferedPagesTotal         int
+	MaxBufferedPagesPerConnection int
+	DebugAddr                     string
 }
 
 func NewConfig() Config {
 	c := Config{
-		Maxcount:          *maxcount,
-		Statsevery:        *statsevery,
-		Lazy:              *lazy,
-		Nodefrag:          *nodefrag,
-		Checksum:          *checksum,
-		Nooptcheck:        *nooptcheck,
-		Ignorefsmerr:      *ignorefsmerr,
-		Allowmissinginit:  *allowmissinginit,
-		Verbose:           *verbose,
-		Debug:             *debug,
-		Quiet:             *quiet,
-		Interface:         *iface,
-		FileName:          *fname,
-		Snaplen:           *snaplen,
-		TsType:            *tstype,
-		Promiscuous:       *promisc,
-		Timeout:           timeout,
-		PacketSource:      *packetSource,
-		BpfFilter:         *bpfFilter,
-		ChannelBufferSize: *channelBufferSize,
+		Maxcount:                      *maxcount,
+		Statsevery:                    *statsevery,
+		Lazy:                          *lazy,
+		Nodefrag:                      *nodefrag,
+		Checksum:                      *checksum,
+		Nooptcheck:                    *nooptcheck,
+		Ignorefsmerr:                  *ignorefsmerr,
+		Allowmissinginit:              *allowmissinginit,
+		Verbose:                       *verbose,
+		Debug:                         *debug,
+		Quiet:                         *quiet,
+		Interface:                     *iface,
+		FileName:                      *fname,
+		Snaplen:                       *snaplen,
+		TsType:                        *tstype,
+		Promiscuous:                   *promisc,
+		StreamFlushTimeout:            time.Duration(*streamFlushTimeout) * time.Second,
+		StreamCloseTimeout:            time.Duration(*streamCloseTimeout) * time.Second,
+		PacketSource:                  *packetSource,
+		BpfFilter:                     *bpfFilter,
+		ChannelBufferSize:             *channelBufferSize,
+		MaxBufferedPagesTotal:         *maxBufferedPagesTotal,
+		MaxBufferedPagesPerConnection: *maxBufferedPagesPerConnection,
+		DebugAddr:                     DebugAddr,
 	}
 
 	// Add filters to only capture common HTTP methods