+++ main.go
@@ -5,14 +5,13 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"runtime"
-	"strconv"
 	"syscall"
 	"time"
 
-	"github.com/honeycombio/ebpf-agent/assemblers"
-	"github.com/honeycombio/ebpf-agent/config"
-	"github.com/honeycombio/ebpf-agent/utils"
+	"github.com/honeycombio/honeycomb-network-agent/assemblers"
+	"github.com/honeycombio/honeycomb-network-agent/config"
+	"github.com/honeycombio/honeycomb-network-agent/debug"
+	"github.com/honeycombio/honeycomb-network-agent/utils"
 	"github.com/honeycombio/libhoney-go"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
@@ -22,30 +21,24 @@ import (
 	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
 )
 
-const Version string = "0.0.10-alpha"
-const defaultDataset = "hny-ebpf-agent"
+const Version string = "0.0.16-alpha"
+const defaultDataset = "hny-network-agent"
 const defaultEndpoint = "https://api.honeycomb.io"
 
 func main() {
 	// Set logging level
 	zerolog.SetGlobalLevel(zerolog.InfoLevel)
-	if os.Getenv("DEBUG") == "true" {
+	if os.Getenv("LOG_LEVEL") == "DEBUG" {
 		zerolog.SetGlobalLevel(zerolog.DebugLevel)
+	}
+	if os.Getenv("DEBUG") == "true" {
+		debug := &debug.DebugService{}
+		debug.Start()
 	}
 	// TODO: add a flag to enable human readable logs
 	// log.Logger = log.Output(zerolog.NewConsoleWriter())
 
-	log.Info().Str("agent_version", Version).Msg("Starting Honeycomb eBPF agent")
-
-	kernelVersion, err := utils.HostKernelVersion()
-	if err != nil {
-		log.Fatal().Err(err).Msg("Failed to get host kernel version")
-	}
-	btfEnabled := utils.HostBtfEnabled()
-	log.Info().
-		Str("kernel_version", kernelVersion.String()).
-		Bool("btf_enabled", btfEnabled).
-		Msg("Detected host kernel")
+	log.Info().Str("agent_version", Version).Msg("Starting Honeycomb Network agent")
 
 	apikey := os.Getenv("HONEYCOMB_API_KEY")
 	if apikey == "" {
@@ -67,12 +60,10 @@ func main() {
 	})
 
 	// appends libhoney's user-agent (TODO: doesn't work, no useragent right now)
-	libhoney.UserAgentAddition = fmt.Sprintf("hny/ebpf-agent/%s", Version)
+	libhoney.UserAgentAddition = fmt.Sprintf("hny/network-agent/%s", Version)
 
 	// configure global fields that are set on all events
 	libhoney.AddField("honeycomb.agent_version", Version)
-	libhoney.AddField("meta.kernel_version", kernelVersion.String())
-	libhoney.AddField("meta.btf_enabled", btfEnabled)
 
 	defer libhoney.Close()
 
@@ -123,13 +114,31 @@ func sendHttpEventToHoneycomb(event assemblers.HttpEvent, k8sClient *utils.Cache
 	// create libhoney event
 	ev := libhoney.NewEvent()
 
+	// calculate event duration
+	// TODO: This is a hack to work around a bug that results in the response timestamp sometimes
+	// being zero which causes the event duration to be negative.
+	if event.RequestTimestamp.IsZero() {
+		log.Debug().
+			Str("request_id", event.RequestId).
+			Msg("Request timestamp is zero")
+		ev.AddField("http.request.timestamp_missing", true)
+		event.RequestTimestamp = time.Now()
+	}
+	if event.ResponseTimestamp.IsZero() {
+		log.Debug().
+			Str("request_id", event.RequestId).
+			Msg("Response timestamp is zero")
+		ev.AddField("http.response.timestamp_missing", true)
+		event.ResponseTimestamp = time.Now()
+	}
+	eventDuration := event.ResponseTimestamp.Sub(event.RequestTimestamp)
+
 	// common attributes
 	ev.Timestamp = event.RequestTimestamp
-	ev.AddField("httpEvent_handled_at", time.Now())
-	ev.AddField("meta.httpEvent_request_handled_latency_ms", time.Now().Sub(event.RequestTimestamp).Milliseconds())
-	ev.AddField("meta.httpEvent_response_handled_latency_ms", time.Now().Sub(event.ResponseTimestamp).Milliseconds())
-	ev.AddField("goroutine_count", runtime.NumGoroutine())
-	ev.AddField("duration_ms", event.Duration.Milliseconds())
+	ev.AddField("meta.httpEvent_handled_at", time.Now())
+	ev.AddField("meta.httpEvent_request_handled_latency_ms", time.Since(event.RequestTimestamp).Milliseconds())
+	ev.AddField("meta.httpEvent_response_handled_latency_ms", time.Since(event.ResponseTimestamp).Milliseconds())
+	ev.AddField("duration_ms", eventDuration.Milliseconds())
 	ev.AddField("http.request.timestamp", event.RequestTimestamp)
 	ev.AddField("http.response.timestamp", event.ResponseTimestamp)
 	ev.AddField("http.request.id", event.RequestId)
@@ -142,14 +151,11 @@ func sendHttpEventToHoneycomb(event assemblers.HttpEvent, k8sClient *utils.Cache
 	// request attributes
 	if event.Request != nil {
 		requestURI = event.Request.RequestURI
-
-		bodySizeString := event.Request.Header.Get("Content-Length")
-		bodySize, _ := strconv.ParseInt(bodySizeString, 10, 64)
 		ev.AddField("name", fmt.Sprintf("HTTP %s", event.Request.Method))
 		ev.AddField(string(semconv.HTTPMethodKey), event.Request.Method)
 		ev.AddField(string(semconv.HTTPURLKey), requestURI)
 		ev.AddField(string(semconv.UserAgentOriginalKey), event.Request.Header.Get("User-Agent"))
-		ev.AddField("http.request.body.size", bodySize)
+		ev.AddField("http.request.body.size", event.Request.ContentLength)
 	} else {
 		ev.AddField("name", "HTTP")
 		ev.AddField("http.request.missing", "no request on this event")
@@ -157,11 +163,8 @@ func sendHttpEventToHoneycomb(event assemblers.HttpEvent, k8sClient *utils.Cache
 
 	// response attributes
 	if event.Response != nil {
-		bodySizeString := event.Response.Header.Get("Content-Length")
-		bodySize, _ := strconv.ParseInt(bodySizeString, 10, 64)
-
 		ev.AddField(string(semconv.HTTPStatusCodeKey), event.Response.StatusCode)
-		ev.AddField("http.response.body.size", bodySize)
+		ev.AddField("http.response.body.size", event.Response.ContentLength)
 
 	} else {
 		ev.AddField("http.response.missing", "no response on this event")