+++ CHANGELOG.md
@@ -1,4 +1,60 @@
-# Honeycomb eBPF Agent changelog
+# Honeycomb Network Agent changelog
+
+## [0.0.16-alpha] - 2023-09-18
+
+### Fixes
+
+- maint: Use debug instead of info for HTTP parse errors (#191) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- fix: Revert 186 and 187 for injecting version during build (#206) | [Jamie Danielson](https://github.com/JamieDanielson)
+- fix: Close HTTP request & response body readers (#195) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- fix: Reuse buffer in stream reader (#200) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+
+## [0.0.15-alpha] - 2023-09-15
+
+### Enhancements
+
+- Refactor readers to not hold byte buffers (#184) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+
+## [0.0.14-alpha] - 2023-09-15
+
+### Maintenance
+
+- maint: Use go mod replace to pull gopacket dependency (#183) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- maint: update RELEASING now that version is computed from a tag (#187) | [Robb Kidd](https://github.com/robbkidd)
+- ci: inject a version at build time based on tags (#186) | [Robb Kidd](https://github.com/robbkidd)
+- maint: fix debug address (#185) | [Jamie Danielson](https://github.com/JamieDanielson)
+- maint: Refactor TCP stream and reader files (#180) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- maint: Rename project to network agent (#172) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+
+## [0.0.13-alpha] - 2023-09-14
+
+### Enhancements
+
+- feat: Record total and active stream counts (#178) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- feat: add debug service (#177) | [Jamie Danielson](https://github.com/JamieDanielson)
+
+### Maintenance
+
+- maint: Log when request/response timestamp is not set & set to time.Now (#179) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- maint: improve Makefile targets (#176) | [Robb Kidd](https://github.com/robbkidd)
+
+## [0.0.12-alpha] - 2023-09-12
+
+### Maintenance
+
+- maint: Try to use less memory (#170) | @robbkidd
+
+## [0.0.11-alpha] - 2023-09-11
+
+### Enhancements
+
+- feat: Separate stream flush and close timeouts (#162) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- feat: Add config options for gopacket max pages and per-conn pages (#160) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+
+### Maintenance
+
+- maint: Replace custom request/response counter with TCP seq & ack counters (#163) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
+- maint: Remove unnecessary fmt.Sprintf usage (#161) | [Mike Goldsmith](https://github.com/MikeGoldsmith)
 
 ## [0.0.10-alpha] - 2023-09-08